LayoutTests/http/tests/misc/script-defer-after-slow-stylesheet.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
<link rel="stylesheet" href="../../js-test-resources/js-test-style.css">
<script src="../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<dvi id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description('This tests that an inline slow stylesheet executes before deferred scripts.');

var jsTestIsAsync = true;

document.addEventListener("DOMContentLoaded", function() {
  debug('DOMContentLoaded');

  finishJSTest();
}, false);
</script>

<script src="http://127.0.0.1:8000/misc/resources/script-debug-body-background.
js" defer="defer"></script>
<link rel="stylesheet" href="http://127.0.0.1:8000/misc/resources/slow-styleshe
et.cgi">
<script>
var successfullyParsed = true;
</script>
<script src="../../js-test-resources/js-test-post.js"></script>
</body>
</html>
