third_party/blink/web_tests/http/tests/misc/script-defer-after-slow-stylesheet-expected.txt

This tests that an inline slow stylesheet executes before deferred scripts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.body)['background-color'] is "rgb(255, 0, 0)"
DOMContentLoaded
PASS successfullyParsed is true

TEST COMPLETE
